Release 1.1.0
This patch updates version and contact information.

Notice that since there have been changes to the code of the library
(src/hwcap.h) a *revision* bump of the library version-info is required [1].
Moreover, since there has been no changes to the interface, i.e.: programs
linked with the previous version do not need to be recompiled nor relinked,
there is no need to update the *current* and *age* fields of the library
version-info.

Since we want to keep the library and package version identical, we update the
package version to 1.1.0.

[1] http://www.gnu.org/software/libtool/manual/html_node/Updating-version-info.html#Updating-version-info